Finished refactoring the proxy-ish StoreMaster::Packages to DataModel::Packages; removed many uses of namespaces as a result. Increased the timeout for getting the Silo service document, appeared under load.  Expanded the exception classes a bit.
The refactoring to DataModel::Packages introduced what I think is an
identity (map) crisis.  Changes to the Pool object read_preferences
are no longer reflected in the Package copies.pool object, even though
they share the same datamapper id (differ in object_id).  This is
unlikely to be a problem in practice, though.